Renames analytics files
